Reorganize auth modules
Moves the top-level `auth` module into the top-level `rest-api` module,
flattening the `auth::rest_api` module into `auth` itself. Also moves
the `auth::oauth` module to the top-level.
