+++ libsplinter/src/oauth/error.rs
