+++ libsplinter/src/oauth/rest_api/resources/callback.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::auth::oauth::UserInfo;
+use crate::oauth::UserInfo;
 
 #[derive(Deserialize)]
 pub struct CallbackQuery {
