+++ libsplinter/src/rest_api/auth/identity/github.rs
