+++ libsplinter/src/rest_api/auth/identity/openid.rs
